reorganizing function handlers
