+++ apps/google-docs/contentful-app-manifest.json
@@ -25,8 +25,8 @@
       "id": "createPlanFunction",
       "name": "Create plan ",
       "description": "Function to create an entry creation plan",
-      "path": "functions/createPlan.js",
-      "entryFile": "functions/createPlan.ts",
+      "path": "functions/handlers/createPlan.js",
+      "entryFile": "functions/handlers/createPlan.ts",
       "allowNetworks": [
         "https://api.openai.com",
         "https://docs.googleapis.com",
@@ -40,8 +40,8 @@
       "id": "createEntriesFromDocumentFunction",
       "name": "Create content entries from document function",
       "description": "Function to create content blocks from App Action.",
-      "path": "functions/createEntriesFromDocument.js",
-      "entryFile": "functions/createEntriesFromDocument.ts",
+      "path": "functions/handlers/createEntriesFromDocument.js",
+      "entryFile": "functions/handlers/createEntriesFromDocument.ts",
       "allowNetworks": [
         "https://api.openai.com",
         "https://docs.googleapis.com",
@@ -56,8 +56,8 @@
       "id": "initiateGdocOauth",
       "name": "Initiate Gdoc OAuth Flow",
       "description": "Initiates the OAuth flow for Google Docs",
-      "path": "functions/initiateOauth.js",
-      "entryFile": "functions/initiateOauth.ts",
+      "path": "functions/oauth/initiateOauth.js",
+      "entryFile": "functions/oauth/initiateOauth.ts",
       "allowNetworks": [
         "oauth2.googleapis.com"
       ],
@@ -69,8 +69,8 @@
       "id": "completeGdocOauth",
       "name": "Complete Gdoc OAuth Flow",
       "description": "Completes the OAuth flow for Google Docs",
-      "path": "functions/completeOauth.js",
-      "entryFile": "functions/completeOauth.ts",
+      "path": "functions/oauth/completeOauth.js",
+      "entryFile": "functions/oauth/completeOauth.ts",
       "allowNetworks": [
         "oauth2.googleapis.com"
       ],
@@ -82,8 +82,8 @@
       "id": "revokeGdocOauthToken",
       "name": "Revoke Gdoc OAuth Token",
       "description": "Revoke token for the Google Docs app to disconnect from the app",
-      "path": "functions/disconnect.js",
-      "entryFile": "functions/disconnect.ts",
+      "path": "functions/oauth/disconnect.js",
+      "entryFile": "functions/oauth/disconnect.ts",
       "allowNetworks": [
         "oauth2.googleapis.com"
       ],
@@ -95,8 +95,8 @@
       "id": "checkGdocOauthTokenStatus",
       "name": "Check Gdoc OAuth Token Status",
       "description": "Checks the status of the Google Docs OAuth token to see if it is valid",
-      "path": "functions/checkStatus.js",
-      "entryFile": "functions/checkStatus.ts",
+      "path": "functions/oauth/checkStatus.js",
+      "entryFile": "functions/oauth/checkStatus.ts",
       "allowNetworks": [
         "oauth2.googleapis.com"
       ],
