+++ apps/google-docs/functions/handlers/createEntriesFromDocument.ts
@@ -4,12 +4,13 @@ import type {
   FunctionTypeEnum,
   AppActionRequest,
 } from '@contentful/node-apps-toolkit';
-import { analyzeContentTypes } from './agents/contentTypeParserAgent/contentTypeParser.agent';
-import { createDocument } from './agents/documentParserAgent/documentParser.agent';
-import { fetchContentTypes } from './service/contentTypeService';
-import { initContentfulManagementClient } from './service/initCMAClient';
-import { fetchGoogleDoc } from './service/googleDriveService';
-import { createEntries } from './service/entryService';
+import { analyzeContentTypes } from '../agents/contentTypeParserAgent/contentTypeParser.agent';
+import { createDocument } from '../agents/documentParserAgent/documentParser.agent';
+import { fetchContentTypes } from '../service/contentTypeService';
+import { initContentfulManagementClient } from '../service/initCMAClient';
+import { fetchGoogleDoc } from '../service/googleDriveService';
+import { createEntries } from '../service/entryService';
+import { parseDocument, validateDocument } from '../utils/documentValidationUtils';
 
 export type AppActionParameters = {
   contentTypeIds: string[];
@@ -34,37 +35,15 @@ export const handler: FunctionEventHandler<
   const { contentTypeIds, document } = event.body;
   const { openAiApiKey } = context.appInstallationParameters as AppInstallationParameters;
 
-  if (!document) {
-    throw new Error('Document is required');
-  }
+  // Validate inputs
+  validateDocument(document);
 
   if (!contentTypeIds || contentTypeIds.length === 0) {
     throw new Error('At least one content type ID is required');
   }
 
-  // Parse document if it's a string (may be JSON stringified during transmission)
-  let parsedDocument: unknown = document;
-  if (typeof document === 'string') {
-    // Check if it's a URL (starts with http:// or https://)
-    if (document.startsWith('http://') || document.startsWith('https://')) {
-      throw new Error(
-        'Document URL provided but fetching from Google Docs API is not yet implemented. Please provide the document JSON object directly.'
-      );
-    }
-
-    // Try to parse as JSON
-    try {
-      parsedDocument = JSON.parse(document);
-    } catch (e) {
-      // Provide more helpful error message
-      const preview = document.substring(0, 100);
-      throw new Error(
-        `Failed to parse document as JSON. Document appears to be a string but is not valid JSON. ` +
-          `Preview: ${preview}${document.length > 100 ? '...' : ''}. ` +
-          `Error: ${e instanceof Error ? e.message : String(e)}`
-      );
-    }
-  }
+  // Parse document using shared utility
+  const parsedDocument = parseDocument(document);
 
   const cma = initContentfulManagementClient(context);
   const contentTypes = await fetchContentTypes(cma, new Set<string>(contentTypeIds));
